+++ src/main/java/com/mary/product_microservice_sharik/config/DataInitializer.java
@@ -7,10 +7,11 @@ import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.ApplicationRunner;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.Random;
+import java.util.concurrent.ThreadLocalRandom;
 
 @RequiredArgsConstructor
 @Component
@@ -24,22 +25,21 @@ public class DataInitializer implements ApplicationRunner{
             return;
         }
 
-        Random random = new Random();
         List<String> categories = Arrays.asList("house", "tea", "sport", "computer", "science");
-        List<Product> products = new ArrayList<>();
+        List<Product> products = new ArrayList<>(30);
 
         for (int i = 0; i < 30; i++) {
             Product product = new Product();
             product.setName("Product " + i);
             product.setImageUrl("https://cdn-icons-png.freepik.com/256/17205/17205369.png");
             product.setDescription("Description " + i);
-            product.setPrice(random.nextInt(1000));
-            product.setAmountLeft(random.nextInt(10));
+            product.setPrice(BigDecimal.valueOf(ThreadLocalRandom.current().nextInt(10000)/100));
+            product.setAmountLeft(ThreadLocalRandom.current().nextInt(15));
             product.setAvailable(true);
 
             List<String> categoriesOfProduct = new ArrayList<>();
             for (String category : categories) {
-                if(random.nextInt(5)>2){
+                if(ThreadLocalRandom.current().nextInt(5)>2){
                     categoriesOfProduct.add(category);
                 }
             }
