+++ src/main/java/com/mary/product_microservice_sharik/model/enumClass/SortProductByEnum.java
@@ -1,0 +1,6 @@
+package com.mary.product_microservice_sharik.model.enumClass;
+
+public enum SortProductByEnum{
+    PRICE,
+    NAME
+}
