--- src/main/java/com/mary/product_microservice_sharik/model/enums/SortProductByEnum.java
@@ -1,6 +1,0 @@
-package com.mary.product_microservice_sharik.model.enums;
-
-public enum SortProductByEnum{
-    PRICE,
-    NAME
-}
